Restructure View/ViewModels into their own distinct folders
