+++ CefSharp.MinimalExample.Wpf/MainWindow.xaml
@@ -1,7 +1,7 @@
 ﻿<Window x:Class="CefSharp.MinimalExample.Wpf.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:main="clr-namespace:CefSharp.MinimalExample.Wpf.Views.Main"
+        xmlns:main="clr-namespace:CefSharp.MinimalExample.Wpf.Views"
         Title="{Binding WebBrowser.Title}"
         WindowState="Maximized">
     <main:MainView/>
