+++ CefSharp.MinimalExample.Wpf/Views/MainView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="CefSharp.MinimalExample.Wpf.Views.Main.MainView"
+﻿<UserControl x:Class="CefSharp.MinimalExample.Wpf.Views.MainView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
