+++ CefSharp.MinimalExample.Wpf/Views/MainView.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System.Windows.Controls;
+using CefSharp.MinimalExample.Wpf.ViewModels;
 
-namespace CefSharp.MinimalExample.Wpf.Views.Main
+namespace CefSharp.MinimalExample.Wpf.Views
 {
     public partial class MainView : UserControl
     {
